feat: configuration 수정
