+++ src/main/java/com/nuzzle/backend/global/config/S3Config.java
@@ -1,27 +1,27 @@
-//package com.nuzzle.backend.global.config;
-//
-//import com.amazonaws.auth.AWSStaticCredentialsProvider;
-//import com.amazonaws.auth.BasicAWSCredentials;
-//import com.amazonaws.services.s3.AmazonS3Client;
-//import com.amazonaws.services.s3.AmazonS3ClientBuilder;
-//import org.springframework.beans.factory.annotation.Value;
-//import org.springframework.context.annotation.Bean;
-//import org.springframework.context.annotation.Configuration;
-//
-//@Configuration
-//public class S3Config {
-//    @Value("${cloud.aws.credentials.access-key}")
-//    private String accessKey;
-//    @Value("${cloud.aws.credentials.secret-key}")
-//    private String secretKey;
-//    @Value("${cloud.aws.region.static}")
-//    private String region;
-//
-//    @Bean
-//    public AmazonS3Client amazonS3Client(){
-//        return (AmazonS3Client) AmazonS3ClientBuilder.standard()
-//                .withRegion(region).enablePathStyleAccess()
-//                .withCredentials(new AWSStaticCredentialsProvider(new BasicAWSCredentials(accessKey, secretKey)))
-//                .build();
-//    }
-//}
+package com.nuzzle.backend.global.config;
+
+import com.amazonaws.auth.AWSStaticCredentialsProvider;
+import com.amazonaws.auth.BasicAWSCredentials;
+import com.amazonaws.services.s3.AmazonS3Client;
+import com.amazonaws.services.s3.AmazonS3ClientBuilder;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class S3Config {
+    @Value("${cloud.aws.credentials.access-key}")
+    private String accessKey;
+    @Value("${cloud.aws.credentials.secret-key}")
+    private String secretKey;
+    @Value("${cloud.aws.region.static}")
+    private String region;
+
+    @Bean
+    public AmazonS3Client amazonS3Client(){
+        return (AmazonS3Client) AmazonS3ClientBuilder.standard()
+                .withRegion(region).enablePathStyleAccess()
+                .withCredentials(new AWSStaticCredentialsProvider(new BasicAWSCredentials(accessKey, secretKey)))
+                .build();
+    }
+}
