--- src/main/java/com/nuzzle/backend/picture/Configuration/S3Config.java
@@ -1,35 +1,0 @@
-package com.nuzzle.backend.picture.Configuration;
-
-import com.amazonaws.auth.AWSStaticCredentialsProvider;
-import com.amazonaws.auth.BasicAWSCredentials;
-import com.amazonaws.services.s3.AmazonS3;
-import com.amazonaws.services.s3.AmazonS3Client;
-import com.amazonaws.services.s3.AmazonS3ClientBuilder;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.beans.factory.annotation.Value;
-
-
-@Configuration
-public class S3Config {
-
-    @Value("${spring.cloud.aws.credentials.accessKey}")
-    private String accessKey;
-
-    @Value("${spring.cloud.aws.credentials.secretKey}")
-    private String secretKey;
-
-    @Value("${spring.cloud.aws.region.static}")
-    private String region;
-
-    @Bean
-    public AmazonS3Client amazonS3Client() {
-        BasicAWSCredentials awsCredentials = new BasicAWSCredentials(accessKey, secretKey);
-
-        return (AmazonS3Client) AmazonS3ClientBuilder.standard()
-//                .withRegion(region)
-                .withRegion("ap-northeast-2")
-                .withCredentials(new AWSStaticCredentialsProvider(awsCredentials))
-                .build();
-    }
-}
